[REFACTOR(Answer)] Answer 클래스 패키지변경
