+++ src/test/kotlin/domain/AnswerTest.kt
@@ -3,6 +3,7 @@ package domain
 import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.ValueSource
+import view.Answer
 
 class AnswerTest {
     @ParameterizedTest
